refactored grid mobile file in components for modularity
